split the repo to interface and it's implementation
