+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/client/remote/ClientRemoteDataSource.kt
@@ -1,0 +1,12 @@
+package com.wire.kalium.logic.data.client.remote
+
+import com.wire.kalium.logic.CoreFailure
+import com.wire.kalium.logic.data.client.Client
+import com.wire.kalium.logic.data.client.RegisterClientParam
+import com.wire.kalium.logic.functional.Either
+
+
+interface ClientRemoteDataSource {
+    suspend fun registerClient(param: RegisterClientParam): Either<CoreFailure, Client>
+
+    }
